% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assert-is-code.R, R/is-code.R
\name{assert_is_valid_r_code}
\alias{assert_is_valid_r_code}
\alias{is_valid_r_code}
\title{Is the input valid R code?}
\usage{
assert_is_valid_r_code(x, severity = getOption("assertive.severity", "stop"))

is_valid_r_code(x, .xname = get_name_in_parent(x))
}
\arguments{
\item{x}{Input to check.}

\item{severity}{How severe should the consequences of the assertion be?
Either \code{"stop"}, \code{"warning"}, \code{"message"}, or \code{"none"}.}

\item{.xname}{Not intended to be used directly.}
}
\value{
\code{TRUE} if the input string is valid R code.
}
\description{
Check to see if the input is a valid (parseable) R code.
}
\examples{
is_valid_r_code("x <- 1 + sqrt(pi)")
is_valid_r_code("x <- ")
is_valid_r_code("<- 1 + sqrt(pi)")
}
\seealso{
\code{\link[base]{parse}}
}

